use goals of achievement

+++ ppo/storage.py
@@ -89,7 +89,7 @@ class RolloutStorage(object):
                     achieved=achieved[_step],
                     params=self.reward_params,
                     dim=1,
-                ).view(next_value.shape)
+                ).view(next_value.shape).float()
 
             # TODO: can we simplify this?
             self.raw_returns = self.raw_returns.detach()
